requirements.txt
pandas
numpy
sentence-transformers
google-genai
tqdm
